… buildSrc/src/main/groovy/com/thoughtworks/go/build/NpmInstallTask.groovy
… gradle/wrapper/gradle-wrapper.jar