## Before

Basically `./gradlew prepare` will take 70 seconds, thanks to a bunch of asset compiles, npm and jruby installs

**Cleanup cache:**

```
rm -rf ~/.gradle/caches/build-cache-1
```

**Warmup:**

```shell
scripts/clean-workingdir.sh; pkill -f java; ./gradlew prepare

....
BUILD SUCCESSFUL in 1m 16s
104 actionable tasks: 102 executed, 2 up-to-date
```

**Run build:**

```shell
scripts/clean-workingdir.sh; pkill -f java; ./gradlew prepare

...
BUILD SUCCESSFUL in 57s
104 actionable tasks: 70 executed, 32 from cache, 2 up-to-date
```

## After

**Cleanup cache:**

```
rm -rf ~/.gradle/caches/build-cache-1
```

**Warmup:**

```shell
scripts/clean-workingdir.sh; pkill -f java; ./gradlew prepare

....
BUILD SUCCESSFUL in 1m 14s
104 actionable tasks: 102 executed, 2 up-to-date
```

**Run build:**

```shell
scripts/clean-workingdir.sh; pkill -f java; ./gradlew prepare;

...
BUILD SUCCESSFUL in 26s
104 actionable tasks: 65 executed, 37 from cache, 2 up-to-date
```