@ketan 2017-08-08
This feels like a hack thanks to how `CacheableExecTask` is implemented. The inputs and outputs of each `CacheableExecTask` still leaks in every place it is used. Unlike `ExtractJRubyTask` where the inputs and outputs are internalized.

This is something I'll cleanup in a later commit.